requirements.txt
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
flasgger==0.9.7.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mistune==3.1.3
mysqlclient==2.2.0
cryptography==41.0.7
packaging==25.0
pluggy==1.5.0
pytest==8.3.5
pytest-flask==1.3.0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.1
Werkzeug==3.1.3
Flask-Mail==0.9.1
pytesseract
pdf2image
Pillow==10.0.0
openai==1.58.1
google-generativeai==0.8.3
